\name{galogistic-package}
\alias{galogistic-package}
\alias{galogistic}
\docType{package}
\title{
What the package does (short line)
Builds a linear model using the genetic algorithm for variable selection
}
\description{
More about what it does (maybe more than one line)
Builds a linear model using the genetic algorithm for variable selection
}
\details{
\tabular{ll}{
Package: \tab galogistic\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-08-06\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Wesley Brooks

Maintainer: Wesley Brooks <wrbrooks@usgs.gov>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
